print to file, not stdout, changed conventions
mainly this update is changing it so that it follows the conventions that he laid out in the assignment sheet. i wasn't sure if he meant he wanted it to look like the first one or the second one, but the second example seemed more likely so i went with that.